Add conditional rule evaluation (and other fixes)
Add the ability for rules to specify what should happen in the event of
failure or success. Three options are available in either case:
- stop program execution (abort),
- continue program execution, or
- skip the next N rules.

Use this new capability to conditionally set wind_field if the 'WI'
variable is not available.

Also, trim airspeed using either TASF or TASG if one is not available.